Add files via upload
This script loops through the simulated STEM images of different crystal structures and orientations and applies a simple rotation of the images for a given rotational increment so far. The augmented images are then cropped so they all have a field of view of ~1.2nm, similar to the window size used for the CNN.

All images and augmented data are stored into .hdf5 file format.

The script so far is for testing purposes and can be extended to also add Poisson noise, image distortions etc., which needs to be discussed.